requirements.txt
asgiref==3.8.1
Django==4.0
django-ckeditor==6.7.1
django-ckeditor-5==0.2.15
django-cors-headers==4.5.0
django-cron==0.6.0
django-environ==0.11.2
django-js-asset==2.2.0
django-markdownx==4.0.7
djangorestframework==3.13.1
Markdown==3.7
pillow==11.0.0
pytz==2024.2
sqlparse==0.5.1
